Add <r4-reset-password>
