+++ src/pages/r4-page-sign.js
@@ -1,42 +1,6 @@
 import page from 'page/page.mjs'
 import { LitElement } from 'lit'
 import {html, literal, unsafeStatic} from 'lit/static-html.js'
-import sdk from '@radio4000/sdk'
-
-export class R4ResetPasswordForm extends LitElement {
-	static properties = {
-		email: {type: String, reflect: true},
-		msg: {type: String, state: true},
-		loading: {type: Boolean, state: true}
-	}
-
-	async submit(event) {
-		event.preventDefault()
-		this.loading = true
-		const email = event.target.email.value
-		if (!email) return
-		const {error} = await sdk.supabase.auth.resetPasswordForEmail(email, {redirectTo: '/settings'})
-		if (error) console.log(error)
-		this.msg = error ? error.message : 'Check your email to continue resetting your password'
-		this.loading = false
-	}
-
-	render() {
-		return html`
-			<form @submit=${this.submit}>
-				<label>Email <input type="email" name="email" value=${this.email} autocomplete="username" required></label>
-				<button type="submit" ?disabled=${this.loading}>${this.loading ? 'Sending...' : 'Send a password reset request'}</button>
-				${this.msg ? html`<p>${this.msg}</p>` : null}
-			</form>
-		`
-	}
-
-	createRenderRoot() {
-		return this
-	}
-}
-
-customElements.define('r4-reset-password', R4ResetPasswordForm)
 
 export default class R4PageSign extends LitElement {
 	static properties = {
